Create dataAccess.java
When you compile and run this file, assuming you have the server running on port 8000, will create a table called "organizations" and insert one hard coded entry into the database. There is currently no operation for deleting items from the database, so running it multiple times in a row without deleting local db files will cause errors. This is only for testing of the database and it will not work like this in the future. In fact, I will probably change this tomorrow before class.